requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
beautifulsoup4==4.12.3
requests==2.31.0
feedgen==1.0.0
python-dateutil==2.9.0.post0
google-genai==1.20.0
python-dotenv==1.0.1 